# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Fresh Platform API
  description: |
    Complete API documentation for the Fresh Platform - a Georgian AI & Technology platform.
    
    ## Authentication
    Most endpoints require JWT authentication via Bearer token in the Authorization header.
    Admin endpoints require a separate admin token via Cookie (`admin_session`).
    
    ## Rate Limiting
    - Search API: 30 requests per minute per IP
    - Admin login: Brute-force protection with lockout
    
    ## Georgian Language
    Many error messages and content are in Georgian.
    
  version: 1.0.0
  contact:
    name: Andrew Altair
    email: andrewaltair@icloud.com
  license:
    name: MIT
servers:
  - url: https://api.example.com
    description: Production server
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/search:
    get:
      tags: [Search]
      summary: Global search
      description: Search across posts, videos, and tools
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
        - name: type
          in: query
          schema:
            type: string
            enum: [posts, videos, tools]
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        type:
                          type: string
                        id:
                          type: string
                        title:
                          type: string
                        description:
                          type: string
                        url:
                          type: string
                        image:
                          type: string
                        category:
                          type: string
                  total:
                    type: integer
                  query:
                    type: string
        '429':
          description: Rate limited

  # ============ TOOLS ============
  /api/tools:
    get:
      tags: [Tools]
      summary: List tools
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: category
          in: query
          schema:
            type: string
        - name: pricing
          in: query
          schema:
            type: string
            enum: [free, freemium, paid]
        - name: featured
          in: query
          schema:
            type: string
            enum: ["true"]
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Tools list
          content:
            application/json:
              schema:
                type: object
                properties:
                  tools:
                    type: array
                    items:
                      $ref: '#/components/schemas/Tool'
                  categories:
                    type: array
                    items:
                      type: string
                  pagination:
                    $ref: '#/components/schemas/Pagination'
    post:
      tags: [Tools]
      summary: Create tool
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ToolInput'
      responses:
        '201':
          description: Tool created

  # ============ VIDEOS ============
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminCookie:
      type: apiKey
      in: cookie
      name: admin_session
  schemas:
...
    Tool:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        url:
          type: string
        logo:
          type: string
        category:
          type: string
        pricing:
          type: string
          enum: [free, freemium, paid]
        rating:
          type: number
        featured:
          type: boolean
        views:
          type: integer
    ToolInput:
      type: object
      required: [name, description, url, category]
      properties:
        name:
          type: string
        description:
          type: string
        url:
          type: string
        logo:
          type: string
        category:
          type: string
        pricing:
          type: string
        rating:
          type: number
        featured:
          type: boolean
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
...